Split i18n/mod.rs into domain submodules
Break the 2100-line monolith into focused files:

- i18n/infra.rs   — infra + monitoring messages
- i18n/service.rs — service add/info messages
- i18n/rotate.rs  — CA rotation messages
- i18n/status.rs  — status + verify messages
- i18n/init.rs    — init, prompt, error, summary,
                     and next-steps messages

mod.rs retains only the Strings struct, Messages
struct, new()/strings(), format_template(), and
tests. No logic or public API changes.

Closes #379
Refs #369
